# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: NovaChat Canonical HTTP API
  version: 1.0.0
  description: |
    Source-controlled contract for the authenticated NovaChat web client.
    Cassandra remains authoritative; cursors are opaque and must not be decoded by clients.
servers:
  - url: http://localhost:8084/api
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /users/search:
    get:
      tags: [Users]
      parameters:
        - {$ref: '#/components/parameters/Limit'}
        - {name: q, in: query, required: true, schema: {type: string, minLength: 2}}
        - {name: cursor, in: query, schema: {type: string}}
      responses: {'200': {description: Public user page, content: {application/json: {schema: {$ref: '#/components/schemas/UserSearchPage'}}}}}
  /users/{userId}:
    get:
      tags: [Users]
      parameters: [{$ref: '#/components/parameters/UserId'}]
      responses: {'200': {description: Public profile, content: {application/json: {schema: {$ref: '#/components/schemas/User'}}}}}
  /users/me:
    patch:
      tags: [Users]
      requestBody:
        content: {application/json: {schema: {$ref: '#/components/schemas/UpdateProfileRequest'}}}
      responses: {'200': {description: Updated profile, content: {application/json: {schema: {$ref: '#/components/schemas/User'}}}}}
  /friends:
    get:
      tags: [Friends]
      parameters: [{$ref: '#/components/parameters/Limit'}]
      responses: {'200': {description: Accepted friendships, content: {application/json: {schema: {$ref: '#/components/schemas/FriendshipStatus'}}}}}
  /friends/request:
    post:
      tags: [Friends]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/FriendRequest'}}}}
      responses: {'200': {description: Pending request}}
...
  /friends/status/{status}:
    get:
      tags: [Friends]
      parameters:
        - {name: status, in: path, required: true, schema: {type: string}}
        - {name: page, in: query, schema: {type: integer, minimum: 0, default: 0}}
        - {$ref: '#/components/parameters/Limit'}
      responses: {'200': {description: Filtered friendship status, content: {application/json: {schema: {$ref: '#/components/schemas/FriendshipStatus'}}}}}
  /friends/accept:
    put:
      tags: [Friends]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/FriendAction'}}}}
      responses: {'200': {description: Accepted friendship}}
  /friends/reject:
    put:
      tags: [Friends]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/FriendAction'}}}}
      responses: {'204': {description: Rejected}}
  /friends/{friendId}/cancel:
    post:
      tags: [Friends]
      parameters: [{$ref: '#/components/parameters/FriendId'}]
      responses: {'204': {description: Cancelled}}
  /friends/{friendId}:
    delete:
      tags: [Friends]
      parameters: [{$ref: '#/components/parameters/FriendId'}]
      responses: {'204': {description: Removed}}
  /friends/block/{friendId}:
    post:
      tags: [Friends]
      parameters: [{$ref: '#/components/parameters/FriendId'}]
      responses: {'204': {description: Blocked}}
  /friends/unblock/{friendId}:
    post:
      tags: [Friends]
      parameters: [{$ref: '#/components/parameters/FriendId'}]
      responses: {'204': {description: Unblocked}}
  /friends/check-block/{friendId}:
    get:
      tags: [Friends]
      parameters: [{$ref: '#/components/parameters/FriendId'}]
      responses: {'200': {description: Blocking relationship}}
  /friends/mutual/{userId}:
    get:
      tags: [Friends]
      parameters:
        - {$ref: '#/components/parameters/UserId'}
        - {name: page, in: query, schema: {type: integer, minimum: 0, default: 0}}
        - {$ref: '#/components/parameters/Limit'}
      responses: {'200': {description: Mutual friends}}
...
  /conversations/{conversationId}/roles/{roleId}:
    delete:
      tags: [Conversations]
      parameters: [{name: conversationId, in: path, required: true, schema: {type: string, format: uuid}}, {name: roleId, in: path, required: true, schema: {type: string, format: uuid}}]
      responses: {'204': {description: Role deleted}}
...
  /polls:
    post:
      tags: [Polls]
      requestBody: {content: {application/json: {schema: {$ref: '#/components/schemas/PollCreate'}}}}
      responses: {'201': {description: Poll created}}
  /polls/{pollId}:
    get:
      tags: [Polls]
      parameters: [{name: pollId, in: path, required: true, schema: {type: string, format: uuid}}]
      responses: {'200': {description: Poll view}}
  /polls/{pollId}/votes:
    post:
      tags: [Polls]
      parameters: [{name: pollId, in: path, required: true, schema: {type: string, format: uuid}}]
      requestBody: {content: {application/json: {schema: {type: object, properties: {selectedOptionIndexes: {type: array, items: {type: integer}}}}}}}
      responses: {'200': {description: Vote recorded}}
    delete:
      tags: [Polls]
      parameters: [{name: pollId, in: path, required: true, schema: {type: string, format: uuid}}]
      responses: {'200': {description: Vote removed}}
  /polls/{pollId}/close:
    post:
      tags: [Polls]
      parameters: [{name: pollId, in: path, required: true, schema: {type: string, format: uuid}}]
      responses: {'200': {description: Poll closed}}
  /invites:
    post:
      tags: [Invites]
      requestBody: {content: {application/json: {schema: {$ref: '#/components/schemas/InviteCreate'}}}}
      responses: {'201': {description: Invite created}}
  /invites/conversation/{conversationId}:
    get:
      tags: [Invites]
      parameters: [{name: conversationId, in: path, required: true, schema: {type: string, format: uuid}}, {$ref: '#/components/parameters/Limit'}]
      responses: {'200': {description: Active invite links}}
  /invites/consume:
    post:
      tags: [Invites]
      requestBody: {content: {application/json: {schema: {$ref: '#/components/schemas/InviteConsume'}}}}
      responses: {'200': {description: Invite consumed}}
  /invites/{token}:
    delete:
      tags: [Invites]
      parameters: [{name: token, in: path, required: true, schema: {type: string}}]
      responses: {'204': {description: Invite revoked}}
  /invites/{token}/decline:
    post:
      tags: [Invites]
      parameters: [{name: token, in: path, required: true, schema: {type: string}}]
      responses: {'200': {description: Invite declined}}
  /invites/conversation/{conversationId}/requests:
    get:
      tags: [Invites]
      parameters: [{name: conversationId, in: path, required: true, schema: {type: string, format: uuid}}, {$ref: '#/components/parameters/Limit'}]
      responses: {'200': {description: Pending join requests}}
  /invites/conversation/{conversationId}/requests/{requestId}/resolve:
    post:
      tags: [Invites]
      parameters: [{name: conversationId, in: path, required: true, schema: {type: string, format: uuid}}, {name: requestId, in: path, required: true, schema: {type: string, format: uuid}}]
      requestBody: {content: {application/json: {schema: {$ref: '#/components/schemas/JoinRequestDecision'}}}}
      responses: {'200': {description: Join request resolved}}
  /public/invites/{token}:
    get:
      tags: [Invites]
      security: []
      parameters: [{name: token, in: path, required: true, schema: {type: string}}]
      responses: {'200': {description: Public preview}}
...
  /admin/users/{userId}/app-roles:
    get:
      tags: [Admin]
      parameters: [{$ref: '#/components/parameters/UserId'}]
      responses: {'200': {description: Application roles}}
    post:
      tags: [Admin]
      parameters: [{$ref: '#/components/parameters/UserId'}]
      requestBody: {content: {application/json: {schema: {$ref: '#/components/schemas/AppRoleMutation'}}}}
      responses: {'201': {description: Application role assigned}}
...
  /admin/users/{userId}/app-roles/{roleCode}:
    delete:
      tags: [Admin]
      parameters: [{name: userId, in: path, required: true, schema: {type: string, format: uuid}}, {name: roleCode, in: path, required: true, schema: {type: string}}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/AppRoleMutation'}}}}
      responses: {'204': {description: Application role revoked}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
...
  parameters:
    UserId: {name: userId, in: path, required: true, schema: {type: string, format: uuid}}
    FriendId: {name: friendId, in: path, required: true, schema: {type: string, format: uuid}}
...
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 200, default: 50}}
  schemas:
...
    User: {type: object, required: [userId, username, accountStatus], properties: {userId: {type: string, format: uuid}, username: {type: string}, email: {type: string, format: email}, displayName: {type: string}, avatarUrl: {type: string, format: uri}, accountStatus: {type: string}}}
    UpdateProfileRequest: {type: object, properties: {displayName: {type: string, maxLength: 120}, avatarUrl: {type: string, format: uri}}}
    UserSearchPage: {type: object, required: [content, hasNext], properties: {content: {type: array, items: {$ref: '#/components/schemas/User'}}, nextCursor: {type: string}, hasNext: {type: boolean}}}
...
    FriendRequest: {type: object, required: [recipientId], properties: {recipientId: {type: string, format: uuid}, message: {type: string, maxLength: 500}}}
    FriendAction: {type: object, required: [friendId], properties: {friendId: {type: string, format: uuid}}}
    FriendUserSummary: {type: object, required: [userId, username, displayName, accountStatus], properties: {userId: {type: string, format: uuid}, username: {type: string}, displayName: {type: string}, avatarUrl: {type: string, format: uri, nullable: true}, accountStatus: {type: string}}}
    FriendshipStatus: {type: object, required: [status, userId, userDetails], properties: {status: {type: string}, userId: {type: string, format: uuid}, userDetails: {type: array, items: {$ref: '#/components/schemas/FriendUserSummary'}}}}
...
    PollCreate: {type: object, required: [conversationId, clientMessageId, question, options], properties: {conversationId: {type: string, format: uuid}, clientMessageId: {type: string, format: uuid}, question: {type: string}, options: {type: array, minItems: 2, maxItems: 10, items: {type: string}}}}
    InviteCreate: {type: object, required: [conversationId], properties: {conversationId: {type: string, format: uuid}, inviteKind: {type: string, enum: [LINK, QR]}, joinPolicy: {type: string, enum: [DIRECT_JOIN, REQUEST_APPROVAL]}, maxUses: {type: integer}, expiresAt: {type: string, format: date-time}}}
    InviteConsume: {type: object, required: [linkToken], properties: {linkToken: {type: string}}}
    JoinRequestDecision: {type: object, required: [requestedAt, decision], properties: {requestedAt: {type: string, format: uuid}, decision: {type: string, enum: [APPROVE, DECLINE]}, reason: {type: string}}}
    AppRoleMutation: {type: object, required: [roleCode, reason], properties: {roleCode: {type: string}, expiresAt: {type: string, format: date-time}, reason: {type: string, minLength: 1, maxLength: 500}}}
...